+++ src/gui/mod.rs
@@ -33,6 +33,7 @@ pub mod stats;
 pub mod theme;
 pub mod tray;
 pub mod tray_status_box;
+pub mod uninstall_dialog;
 pub mod update_prompt;
 pub mod window_state;